feat: read job links file if present

+++ Wuzzuf_DataCollection/__main__.py
@@ -1,4 +1,5 @@
 import json
+import os.path
 import time
 from pathlib import Path
 
@@ -6,14 +7,20 @@ from Wuzzuf_DataCollection.getjobinfo import getJobInfo
 from Wuzzuf_DataCollection.getjoblinks import getJobLinks
 
 OUTPUT_FOLDER = "output"
+OUTPUT_LINKS_FILE = f"{OUTPUT_FOLDER}/links.json"
 TIME_BETWEEN_REQUESTS = 2.5
 
 Path(OUTPUT_FOLDER).mkdir(parents=True, exist_ok=True)
 
-jobLinks = getJobLinks()
+if os.path.isfile(OUTPUT_LINKS_FILE):
+    print("Links file found")
+    with open(OUTPUT_LINKS_FILE, "r") as outputFile:
+        jobLinks = json.load(outputFile)
+else:
+    jobLinks = getJobLinks()
 
-with open(f"{OUTPUT_FOLDER}/links.json", "w") as outputFile:
-    json.dump(jobLinks, outputFile)
+    with open(OUTPUT_LINKS_FILE, "w") as outputFile:
+        json.dump(jobLinks, outputFile)
 
 
 for i, link in enumerate(jobLinks):
